# api/swagger.yaml
openapi: 3.0.3
info:
    title: URLAsker API
    version: 0.0.1
servers:
    - url: 'http://localhost'
paths:
    /api/v1/info:
        get:
            summary: Get info about asker
            operationId: info
            responses:
                '200':
                    description: Success
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    interval:
                                        type: string
                                    urls:
                                        type: array
                                example:
                                    interval: '5s'
                                    urls: ["https://www.google.com/", "https://yandex.com/", "https://www.yahoo.com/", "http://xydsffew.com/"]
                '500':
                    description: internal server error
    /api/v1/listLatest:
        get:
            summary: Get latest result of asker work
            operationId: listLatest
            responses:
                '200':
                    description: Success
                    content:
                        application/json:
                            schema:

                '500':
                    description: internal server error
    /api/v1/list:
    /api/v1/setInterval:
        post:
            summary: Set interval
            operationId: setInterval
            parameters:
                - name: interval
                  in: query
                  schema:
                      type: string
                      example: 15s
                  required: true
            responses:
                '200':
                    description: Success
                    content:
                        text/plain:
                            example: Success
                '400':
                    description: invalid interval param
                '500':
                    description: internal server error
    /api/v1/setUrls:
        post:
            summary: Rewrite urls
            operationId: setUrls
            requestBody:
                content:
                    application/json:
                        schema:
                            type: array
                        example: ["https://www.google.com/", "https://yandex.com/", "https://www.yahoo.com/", "http://xydsffew.com/"]
            responses:
                '200':
                    description: Success
                    content:
                        text/plain:
                            example: Success
                '400':
                    description: invalid body or invalid url in body
                '500':
                    description: internal server error
    /api/v1/addUrls:
        put:
            summary: Append urls
            operationId: addUrls
            requestBody:
                content:
                    application/json:
                        schema:
                            type: array
                        example: ["https://www.google.com/", "https://www.rambler.ru/"]
            responses:
                '200':
                    description: Success
                    content:
                        text/plain:
                            example: Success
                '400':
                    description: invalid body or invalid url in body
                '500':
                    description: internal server error
    /api/v1/deleteUrls:
        delete:
            summary: Delete urls
            operationId: deleteUrls
            requestBody:
                content:
                    application/json:
                        schema:
                            type: array
                        example: ["https://www.google.com/"]
            responses:
                '200':
                    description: Success
                    content:
                        text/plain:
                            example: Success
                '400':
                    description: invalid body or invalid url in body
                '500':
                    description: internal server error